# task-card-cards/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/cards": {
            "get": {
                "description": "Get all cards",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Get all cards",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User id of the card",
                        "name": "user_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Card"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/globals.BaseRequestReturn"
                        }
                    }
                }
            },
            "post": {
                "description": "Route to update a card",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Create a card",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title of the card",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Summary of the card",
                        "name": "summary",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Due date of the card",
                        "name": "due_date",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Status of the card",
                        "name": "card_status",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User id of the card",
                        "name": "user_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/globals.BaseRequestReturn"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/globals.BaseRequestReturn"
                        }
                    }
                }
            }
        },
        "/cards/{id}": {
            "get": {
                "description": "Get a card",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Get a card",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Card id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Card"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/globals.BaseRequestReturn"
                        }
                    }
                }
            },
            "post": {
                "description": "Move a card",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Move a card",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Card id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Status of the card",
                        "name": "card_status",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/globals.BaseRequestReturn"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/globals.BaseRequestReturn"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a card",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Delete a card",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Card id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User id of the card",
                        "name": "user_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/globals.BaseRequestReturn"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/globals.BaseRequestReturn"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update a card",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Update a card",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Card id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Title of the card",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Summary of the card",
                        "name": "summary",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Due date of the card",
                        "name": "due_date",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Status of the card",
                        "name": "card_status",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User id of the card",
                        "name": "user_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/globals.BaseRequestReturn"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/globals.BaseRequestReturn"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "globals.BaseRequestReturn": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.Card": {
            "type": "object",
            "properties": {
                "card_status": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "summary": {
                    "type": "string",
                    "maxLength": 2500
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    }
}